[PATCH] pass SEQNUM trough udevd
here we change udevd to pass the SEQNUM from the hotplug environment
to udev and the dev.d/ scripts. We need this for HAL to match the
hotplug event with the dev.d/ events.

It also changes the type from int to long to match the kernel.

+++ udevd.h
@@ -33,7 +33,7 @@ struct hotplug_msg {
 	char magic[20];
 	struct list_head list;
 	pid_t pid;
-	int seqnum;
+	long seqnum;
 	long queue_time;
 	char action[ACTION_SIZE];
 	char devpath[DEVPATH_SIZE];
